docs: 메인 페이지, 수정 페이지 원위치
